+++ src/main/logutils.ts
@@ -45,6 +45,12 @@ let isClassicRunning: boolean = false;
 let isRecordingChallengeMode: boolean = false;
 
 /**
+ * Timers for poll
+ */
+let pollWowProcessInterval: NodeJS.Timer;
+let watchLogsInterval: NodeJS.Timer;
+ 
+/**
  * wowProcessStarted
  */
 const wowProcessStarted = () => {
@@ -263,6 +269,9 @@ const determineArenaMatchResult = (line: string): any[] => {
  const handleEncounterStopLine = (line: string) => {
     if (!recorder.isRecording) return; 
     
+    // ignore encounter events inside Challenge Modes
+    if (isRecordingChallengeMode) return; 
+
     // ignore encounter events inside Challenge Modes
     if (isRecordingChallengeMode) return; 
 
@@ -489,7 +498,9 @@ const isUnitSelf = (srcFlags: number): boolean => {
 const watchLogs = (logdir: any) => {
     const checkInterval: number = 1000;
     
-    setInterval(() => {
+    if (watchLogsInterval) clearInterval(watchLogsInterval);
+
+    watchLogsInterval = setInterval(() => {
         currentLogFile = getLatestLog(logdir);
         const logFileChanged = (lastLogFile !== currentLogFile)
 
@@ -549,7 +560,10 @@ const checkWoWProcess = async (): Promise<[boolean, boolean]> => {
  * pollWoWProcess
  */
 const pollWowProcess = () => {
-    setInterval(async () => {
+
+    if (pollWowProcessInterval) clearInterval(pollWowProcessInterval);
+
+    pollWowProcessInterval = setInterval(async () => {
         const [retailFound, classicFound] = await checkWoWProcess();
         const retailProcessChanged = (retailFound !== isRetailRunning);    
         // TODO classic support
@@ -559,9 +573,9 @@ const pollWowProcess = () => {
         if (!retailProcessChanged) return;
           
         if (retailFound) {
-            wowProcessStarted(true);
+            wowProcessStarted();
         } else {
-            wowProcessStopped(true);
+            wowProcessStopped();
         }
     }, 5000);
 }